pyproject.toml
[tool.poetry]
name = "aws-cdk-python-demo"
version = "0.1.0"
description = ""
authors = ["Yaroslav Muravskyi <yaroslav.muravskyi@myarik.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
jinja2 = "^3.1.4"
aws-lambda-powertools = {extras = ["tracer"],version = "^3.0.0"}
pydantic = {extras = ["email"],version = "^2.9.2"}

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
aws-cdk-lib = "^2.159.1"
ruff = "^0.6.6"
pytest = "^8.3.3"
aws-cdk-aws-lambda-python-alpha = "^2.159.1a0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
